requirements.txt
streamlit
mysql-connector-python
reportlab
requests
